Added `wharf vars sub` command (#110)
* Moved `wharf vars` to `wharf varsub list`

* Added `wharf vars sub` command

* Added ability to use file instead of STDIN

* Updated CHANGELOG.md
